+++ bugsnag-plugin-android-ndk/src/main/jni/bugsnag.c
@@ -84,7 +84,7 @@ void bsg_populate_notify_stacktrace(JNIEnv *env, bugsnag_stackframe *stacktrace,
     jstring filename = bsg_safe_new_string_utf(env, frame.filename);
     jstring method;
     if (strlen(frame.method) == 0) {
-      char *frame_address = malloc(sizeof(char) * 32);
+      char *frame_address = calloc(1, sizeof(char) * 32);
       sprintf(frame_address, "0x%lx", (unsigned long)frame.frame_address);
       method = bsg_safe_new_string_utf(env, frame_address);
       free(frame_address);